requirements.txt
numpy 
pandas
openpyxl
pyxlsb
snowflake-connector-python
snowflake-snowpark-python
snowflake-sqlalchemy
streamlit
streamlit-authenticator
sqlalchemy
openpyexcel
azure-storage-blob
xlsxwriter
python-dateutil
azure-identity
azure-keyvault
streamlit-modal
streamlit-javascript